chore(authentication): rename authentication middleware
